requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
requests==2.31.0
gspread==5.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
